// packages/wallet/core/src/relayer/standard/rpc/relayer.gen.ts
export enum FeeTokenType {
  UNKNOWN = 'UNKNOWN',
  ERC20_TOKEN = 'ERC20_TOKEN',
  ERC1155_TOKEN = 'ERC1155_TOKEN',
}
// ...
export interface MetaTxn {
  walletAddress: string
  contract: string
  input: string
}